+++ src/diffusers/models/unets/unet_kandinsky3.py
@@ -19,11 +19,11 @@ import torch
 import torch.utils.checkpoint
 from torch import nn
 
-from ..configuration_utils import ConfigMixin, register_to_config
-from ..utils import BaseOutput, logging
-from .attention_processor import Attention, AttentionProcessor, AttnProcessor
-from .embeddings import TimestepEmbedding, Timesteps
-from .modeling_utils import ModelMixin
+from ...configuration_utils import ConfigMixin, register_to_config
+from ...utils import BaseOutput, logging
+from ..attention_processor import Attention, AttentionProcessor, AttnProcessor
+from ..embeddings import TimestepEmbedding, Timesteps
+from ..modeling_utils import ModelMixin
 
 
 logger = logging.get_logger(__name__)  # pylint: disable=invalid-name
