+++ tests/models/test_unet_2d_blocks.py
@@ -14,7 +14,7 @@
 # limitations under the License.
 import unittest
 
-from diffusers.models.unet_2d_blocks import *  # noqa F403
+from diffusers.models.unets.unet_2d_blocks import *  # noqa F403
 from diffusers.utils.testing_utils import torch_device
 
 from .test_unet_blocks_common import UNetBlockTesterMixin
